Split source into diffrent files, so the functionality can be reused in other node applications
